chore: rename some files to *.ts/*.tsx
